requirements.txt
beautifulsoup4==4.10.0
bs4==0.0.1
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.7
colorama==0.4.4
commonmark==0.9.1
google-api-core==2.3.0
google-api-python-client==2.33.0
google-auth==2.3.3
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
googleapis-common-protos==1.54.0
httplib2==0.20.2
idna==3.3
isort==4.3.21
Jinja2==3.0.3
lazy-object-proxy==1.4.3
MarkupSafe==2.0.1
mccabe==0.6.1
oauthlib==3.1.1
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pygame==1.9.6
Pygments==2.10.0
pyparsing==3.0.6
requests==2.26.0
requests-oauthlib==1.3.0
rich==10.12.0
rope==0.17.0
rsa==4.8
six==1.16.0
soupsieve==2.2.1
tqdm==4.62.3
uritemplate==4.1.1
urllib3==1.26.7
wincertstore==0.2
wrapt==1.11.2
